requirements.txt
eth-brownie>=1.11.0,<2.0.0
tox==3.15.1
pytest
requests